requirements.txt
flask
Flask-API
requests
pyserial
requests
